gallery
Some work on the art gallery.
