+++ components/gallery/detail/artwork.js
@@ -1,0 +1,13 @@
+import Image from "next/image"
+
+import styles from '../../../styles/gallery/detail/artwork.module.css'
+
+function Artwork() {
+    return (
+        <div>
+
+        </div>
+    )
+}
+
+export default Artwork
