+++ components/gallery/detail/title.js
@@ -1,0 +1,18 @@
+import Link from "next/link";
+
+import styles from '../../../styles/gallery/detail/artwork.module.css'
+
+function Title() {
+    return (
+        <div className={styles.container}>
+            <Link href="/gallery/">
+                <a className={styles.home}>&lt; Back to Home</a>
+            </Link>
+            <h1>Art Gallery</h1>
+            <p>Description</p>
+            <input className={styles.searchBar} placeholder="Search by artist name or title"></input>
+        </div>
+    )
+}
+
+export default Title
