--- components/gallery/title.js
@@ -1,12 +1,0 @@
-import styles from '../../styles/gallery/title.module.css'
-
-function Title() {
-    return (
-        <div>
-            <h1 className={styles.title}>Art Gallery</h1>
-            <p className={styles.description}>Here are some artworks from the quarter.</p>
-        </div>
-    )
-}
-
-export default Title
