+++ pages/gallery/index.js
@@ -1,6 +1,6 @@
-import Title from '../../components/gallery/title'
-import Artwork from '../../components/gallery/artwork'
-import DetailButton from '../../components/gallery/detail_button'
+import Title from '../../components/gallery/index/title'
+import Artwork from '../../components/gallery/index/artwork'
+import DetailButton from '../../components/gallery/index/detail_button'
 
 import styles from '../../styles/gallery/index.module.css'
