+++ styles/gallery/detail/description.module.css
@@ -1,0 +1,30 @@
+.container {
+    display: flex;
+    justify-content: center;
+    background: #F5EFF5;
+    width: 20%;
+    padding-top: 10em;
+    padding-left: 3em;
+    padding-right: 3em;
+}
+
+.description {
+    width: 100%;
+    word-wrap: break-word;
+}
+
+.navigation {
+    display: flex;
+    width: 100%;
+    justify-content: space-between;
+    flex-direction: row;
+}
+
+.button {
+    border: none;
+    background: none;
+}
+
+.button:hover {
+    text-decoration: underline;
+}
