+++ styles/gallery/detail/title.module.css
@@ -1,0 +1,12 @@
+.container {
+    display: flex;
+    flex-direction: column;
+}
+
+.home:hover {
+    text-decoration: underline;
+}
+
+.searchBar {
+    border: none;
+}
