--- styles/gallery/detail_button.module.css
@@ -1,30 +1,0 @@
-.button {
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    align-items: center;
-    padding: 10px 20px;
-
-    position: static;
-    width: 145px;
-    height: 43px;
-    left: 0px;
-    top: 0px;
-
-    border: 1px solid #232020;
-    box-sizing: border-box;
-    border-radius: 14px;
-
-    flex: none;
-    order: 0;
-    flex-grow: 0;
-    margin: 0px 40px;
-    
-    font-family: Catamaran;
-    font-size: 16px;
-    font-style: normal;
-    font-weight: 400;
-    line-height: 23px;
-    letter-spacing: -0.04em;
-    text-align: center;
-}
