+++ styles/gallery/index/artwork.module.css
@@ -1,0 +1,47 @@
+.container {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    align-items: center;
+
+    position: static;
+    width: 60%;
+    height: 445px;
+
+    flex: none;
+    order: 1;
+    flex-grow: 0;
+    margin: 64px 0px;
+}
+
+.panel {
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+
+    position: static;
+    width: 273px;
+    height: 445px;
+
+    flex: none;
+    order: 2;
+    flex-grow: 0;
+}
+
+.description {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: flex-start;
+    padding: 0px 0px 0px 24px;
+
+    position: static;
+    width: 273px;
+    height: 129px;
+
+    background: #D4B7DF;
+
+    flex: none;
+    order: 1;
+    flex-grow: 0;
+}
