+++ styles/gallery/index/title.module.css
@@ -1,0 +1,17 @@
+.title {
+    font-size: 45px;
+    font-style: normal;
+    font-weight: 600;
+    line-height: 52px;
+    letter-spacing: -0.04em;
+    text-align: center;
+}
+
+.description {
+    font-size: 16px;
+    font-style: normal;
+    font-weight: 400;
+    line-height: 23px;
+    letter-spacing: -0.04em;
+    text-align: center;
+}
