+++ styles/globals.css
@@ -1,9 +1,9 @@
 html,
 body {
   padding: 0;
-  margin: 0;
-  font-family: -apple-system, BlinkMacSystemFont, Segoe UI, Roboto, Oxygen,
-    Ubuntu, Cantarell, Fira Sans, Droid Sans, Helvetica Neue, sans-serif;
+  margin: 2em;
+  background: #FCF7F8;
+  font-family: Catamaran;
 }
 
 a {
